Switched nearly everything over to the new logging system.
