+++ uchicagoldrtoolsuite/bit_level/app/archiver.py
@@ -1,10 +1,10 @@
 from os.path import join, dirname, expanduser, expandvars
+from logging import getLogger
 
 from pypairtree.utils import identifier_to_path
 
 from uchicagoldrtoolsuite.core.app.abc.cliapp import CLIApp
 from uchicagoldrtoolsuite.core.lib.masterlog import \
-    spawn_logger, \
     activate_master_log_file, \
     activate_job_log_file, \
     activate_stdout_log
@@ -22,7 +22,7 @@ __publication__ = ""
 __version__ = "0.0.1dev"
 
 
-log = spawn_logger(__name__)
+log = getLogger(__name__)
 activate_master_log_file()
 activate_job_log_file()
