+++ uchicagoldrtoolsuite/bit_level/app/piecemealstager.py
@@ -1,9 +1,9 @@
 from os.path import join, dirname, expanduser, expandvars
+from logging import getLogger
 
 from uchicagoldrtoolsuite.core.app.abc.cliapp import CLIApp
 from uchicagoldrtoolsuite.core.lib.convenience import recursive_scandir
-from uchicagoldrtoolsuite.core.lib.masterlog import \
-    spawn_logger, \
+from uchicagoldrtoolsuite import \
     activate_master_log_file, \
     activate_job_log_file, \
     activate_stdout_log
@@ -21,7 +21,7 @@ __publication__ = ""
 __version__ = "0.0.1dev"
 
 
-log = spawn_logger(__name__)
+log = getLogger(__name__)
 activate_master_log_file()
 activate_job_log_file()
