+++ uchicagoldrtoolsuite/bit_level/app/presformcreator.py
@@ -1,10 +1,10 @@
 from sys import stdout
+from logging import getLogger
 from os.path import join
 from configparser import NoOptionError
 
 from uchicagoldrtoolsuite.core.app.abc.cliapp import CLIApp
-from uchicagoldrtoolsuite.core.lib.masterlog import \
-    spawn_logger, \
+from uchicagoldrtoolsuite import \
     activate_master_log_file, \
     activate_job_log_file, \
     activate_stdout_log
@@ -22,7 +22,7 @@ __publication__ = ""
 __version__ = "0.0.1dev"
 
 
-log = spawn_logger(__name__)
+log = getLogger(__name__)
 activate_master_log_file()
 activate_job_log_file()
