+++ uchicagoldrtoolsuite/bit_level/app/technicalmetadatacreator.py
@@ -1,4 +1,5 @@
 from sys import stdout
+from logging imnport getLogger
 from os.path import join
 from configparser import NoOptionError
 
@@ -9,8 +10,7 @@ from ..lib.processors.generictechnicalmetadatacreator import \
     GenericTechnicalMetadataCreator
 from ..lib.techmdcreators.fitscreator import FITsCreator
 from ..lib.techmdcreators.apifitscreator import APIFITsCreator
-from uchicagoldrtoolsuite.core.lib.masterlog import \
-    spawn_logger, \
+from uchicagoldrtoolsuite import \
     activate_master_log_file, \
     activate_job_log_file, \
     activate_stdout_log
@@ -23,7 +23,7 @@ __copyright__ = "Copyright University of Chicago, 2016"
 __publication__ = ""
 __version__ = "0.0.1dev"
 
-log = spawn_logger(__name__)
+log = getLogger(__name__)
 activate_master_log_file()
 activate_job_log_file()
