+++ uchicagoldrtoolsuite/bit_level/lib/converters/imageconverter.py
@@ -1,9 +1,9 @@
 from json import dumps
 from os.path import join, isfile
 from uuid import uuid4
+from logging import getLogger
 
 from uchicagoldrtoolsuite.core.lib.bash_cmd import BashCommand
-from uchicagoldrtoolsuite.core.lib.masterlog import spawn_logger
 from .abc.converter import Converter
 
 
@@ -15,7 +15,7 @@ __publication__ = ""
 __version__ = "0.0.1dev"
 
 
-log = spawn_logger(__name__)
+log = getLogger(__name__)
 
 
 class ImageConverter(Converter):
