+++ uchicagoldrtoolsuite/bit_level/lib/converters/officetocsvconverter.py
@@ -2,13 +2,14 @@ from os import scandir, makedirs
 from json import dumps
 from os.path import join, isfile
 from uuid import uuid4
+from logging import getLogger
 
 from uchicagoldrtoolsuite.core.lib.bash_cmd import BashCommand
 from uchicagoldrtoolsuite.core.lib.masterlog import spawn_logger
 from .abc.converter import Converter
 
 
-log = spawn_logger(__name__)
+log = getLogger(__name__)
 
 
 __author__ = "Brian Balsamo"
