+++ uchicagoldrtoolsuite/bit_level/lib/converters/officetopdfconverter.py
@@ -2,9 +2,9 @@ from os import scandir, makedirs
 from json import dumps
 from os.path import join, isfile
 from uuid import uuid4
+from logging import getLogger
 
 from uchicagoldrtoolsuite.core.lib.bash_cmd import BashCommand
-from uchicagoldrtoolsuite.core.lib.masterlog import spawn_logger
 from .abc.converter import Converter
 
 
@@ -16,7 +16,7 @@ __publication__ = ""
 __version__ = "0.0.1dev"
 
 
-log = spawn_logger(__name__)
+log = getLogger(__name__)
 
 
 class OfficeToPDFConverter(Converter):
