+++ uchicagoldrtoolsuite/bit_level/lib/ldritems/ldrpath.py
@@ -1,8 +1,8 @@
 from os import remove
 from pathlib import Path
 from json import dumps
+from logging import getLogger
 
-from uchicagoldrtoolsuite.core.lib.masterlog import spawn_logger
 from .abc.ldritem import LDRItem
 
 
@@ -14,7 +14,7 @@ __publication__ = ""
 __version__ = "0.0.1dev"
 
 
-log = spawn_logger(__name__)
+log = getLogger(__name__)
 
 
 class LDRPath(LDRItem):
