+++ uchicagoldrtoolsuite/bit_level/lib/misc/premisextensionnodes.py
@@ -1,3 +1,5 @@
+from logging import getLogger
+
 from pypremis.nodes import ExtensionNode
 from controlledvocab.lib import ControlledVocabulary as CV
 
@@ -8,6 +10,9 @@ __company__ = "The University of Chicago Library"
 __copyright__ = "Copyright University of Chicago, 2016"
 __publication__ = ""
 __version__ = "0.0.1dev"
+
+
+log = getLogger(__name__)
 
 
 class linkingRightsExtensionIdentifier(ExtensionNode):
