+++ uchicagoldrtoolsuite/bit_level/lib/techmdcreators/abc/technicalmetadatacreator.py
@@ -1,4 +1,5 @@
 from os.path import join
+from logging import getLogger
 from abc import ABCMeta, abstractmethod
 from uuid import uuid1
 
@@ -17,6 +18,9 @@ __company__ = "The University of Chicago Library"
 __copyright__ = "Copyright University of Chicago, 2016"
 __publication__ = ""
 __version__ = "0.0.1dev"
+
+
+log = getLogger(__name__)
 
 
 class TechnicalMetadataCreator(metaclass=ABCMeta):
