+++ uchicagoldrtoolsuite/bit_level/lib/transformers/abc/transformer.py
@@ -1,5 +1,5 @@
-
 from abc import ABCMeta, abstractproperty, abstractmethod
+from logging import getLogger
 
 __author__ = "Tyler Danstrom"
 __email__ = " tdanstrom@uchicago.edu"
@@ -7,6 +7,9 @@ __company__ = "The University of Chicago Library"
 __copyright__ = "Copyright University of Chicago, 2016"
 __publication__ = ""
 __version__ = "0.0.1dev"
+
+
+log = getLogger(__name__)
 
 
 class Transformer(metaclass=ABCMeta):
