+++ uchicagoldrtoolsuite/bit_level/lib/writers/abc/stageserializationwriter.py
@@ -1,4 +1,5 @@
 from abc import ABCMeta, abstractmethod
+from logging import getLogger
 
 from .abc.serializationwriter import SerializationWriter
 from ...structures.stage import Stage
@@ -10,6 +11,9 @@ __company__ = "The University of Chicago Library"
 __copyright__ = "Copyright University of Chicago, 2016"
 __publication__ = ""
 __version__ = "0.0.1dev"
+
+
+log = getLogger(__name__)
 
 
 class StageSerializationWriter(SerializationWriter, metaclass=ABCMeta):
