+++ uchicagoldrtoolsuite/core/lib/bash_cmd.py
@@ -1,4 +1,5 @@
 from subprocess import run, PIPE, STDOUT, TimeoutExpired
+from logging import getLogger
 
 
 __author__ = "Brian Balsamo, Tyler Danstrom"
@@ -7,6 +8,9 @@ __company__ = "The University of Chicago Library"
 __copyright__ = "Copyright University of Chicago, 2016"
 __publication__ = ""
 __version__ = "0.0.1dev"
+
+
+log = getLogger(__name__)
 
 
 class BashCommand(object):
